shop delete function tested and completed

+++ controller/delete_menu.php
@@ -1,10 +1,18 @@
 <?php
-include_once '../convig.php';
-$id = filter_input(INPUT_GET, "n");
-$menu = Menu::getInstance();
 
-if ($menu->delete($id)) {
-     header("location: ../admin/menu?info=del_success");
+include_once '../convig.php';
+if (isset($_GET['n'])) {
+    $id = filter_input(INPUT_GET, "n");
+    $menu = Menu::getInstance();
+
+    if ($menu->delete($id)) {
+        header("location: ../admin/menu?info=del_success");
+        exit;
+    } else {
+        header("location: ../admin/menu?info=invalid");
+        exit;
+    }
 } else {
-     header("location: ../admin/menu?info=invalid");
+    header("location: ../admin/menu?info=invalid");
+    exit;
 }
